+++ examples/09-file-exclusion/config-exclude/mdite.config.js
@@ -1,0 +1,7 @@
+export default {
+  exclude: [
+    'drafts/**',
+    '*.temp.md',
+    'scratch/'
+  ]
+};